requirements.txt
# ========================================
# CORE DATA PROCESSING
# ========================================
polars>=0.20.0              # Fast DataFrame library (principal)
duckdb>=0.10.0              # SQL engine for analytics
pyarrow>=15.0.0             # Arrow format, Parquet I/O
pandas>=2.1.0               # Legacy DataFrame (compatibility)
numpy>=1.24.0               # Numerical computing

# ========================================
# API & HTTP
# ========================================
requests>=2.31.0            # HTTP requests (Polygon API)
urllib3>=2.0.0              # HTTP client library
certifi>=2023.0.0           # SSL certificates (fix TLS Windows)

# ========================================
# MACHINE LEARNING & MODELING
# ========================================
lightgbm>=4.1.0             # Gradient boosting framework
xgboost>=2.0.0              # Extreme gradient boosting
catboost>=1.2.0             # Categorical boosting
scikit-learn>=1.3.0         # ML utilities, metrics
torch>=2.1.0                # Deep learning framework
transformers>=4.35.0        # HuggingFace models
scipy>=1.11.0               # Scientific computing (spearmanr, entropy)

# ========================================
# BACKTESTING
# ========================================
backtrader>=1.9.78          # Backtesting framework
vectorbt>=0.26.0            # Vectorized backtesting

# ========================================
# TECHNICAL ANALYSIS
# ========================================
ta-lib>=0.4.28              # Technical indicators
# pandas-ta  # Skip for now (Python 3.13 compatibility issues)

# ========================================
# VISUALIZATION
# ========================================
matplotlib>=3.8.0           # Plotting library
seaborn>=0.13.0             # Statistical visualization
plotly>=5.18.0              # Interactive plots

# ========================================
# EXPERIMENT TRACKING
# ========================================
mlflow>=2.9.0               # ML experiment tracking

# ========================================
# CONFIGURATION & ENV
# ========================================
pyyaml>=6.0.1               # YAML config files
python-dotenv>=1.0.0        # .env file loader
argparse>=1.4.0             # CLI argument parsing (built-in, pero explicit)

# ========================================
# LOGGING
# ========================================
loguru>=0.7.2               # Advanced logging

# ========================================
# TESTING
# ========================================
pytest>=7.4.0               # Testing framework
pytest-cov>=4.1.0           # Coverage reports

# ========================================
# TYPE HINTS & LINTING
# ========================================
mypy>=1.7.0                 # Static type checker

# ========================================
# CODE QUALITY
# ========================================
black>=23.11.0              # Code formatter
ruff>=0.1.6                 # Fast Python linter

# ========================================
# JUPYTER & NOTEBOOKS
# ========================================
jupyter>=1.0.0              # Jupyter notebook
jupyterlab>=4.0.0           # JupyterLab interface
ipykernel>=6.26.0           # IPython kernel
ipywidgets>=8.1.0           # Interactive widgets

# ========================================
# UTILITIES
# ========================================
tqdm>=4.66.0                # Progress bars
python-dateutil>=2.8.2      # Date utilities (relativedelta)
pytz>=2023.3                # Timezone handling
pathlib>=1.0.1              # Path manipulation (built-in, pero explicit)
hashlib>=20081119           # Hashing (built-in, pero explicit)

# ========================================
# CONCURRENCY & PARALLELIZATION
# ========================================
# concurrent.futures (built-in)
# multiprocessing (built-in)

# ========================================
# DATA FORMATS
# ========================================
# json (built-in)
# csv (built-in)
# zipfile (built-in)

# ========================================
# OPTIONAL: PERFORMANCE MONITORING
# ========================================
# memory_profiler>=0.61.0   # Memory usage profiling (descomentar si necesitas)
# py-spy>=0.3.14            # Sampling profiler (descomentar si necesitas)
